lastfm: first commit
This marks the beginning of an effort to merge my clusterfuck of Lastfm
scripts into a monolith. Data structures and code are a huge mess,
deliberately left in there for my future self to clean up.

At this point, only tracks from a DAP (e.g. running Rockbox) can be
scrobbled. I will (re-)introduce more functionality in the future.